refactored backend api, added authentication to all endpoints, updated error messages on authorization for frontend
